Split test module into two
